feat(arch-w7): 架构 2.0 守护收口——下载线程入监督器（D-77）+ check_deps/check_guards 进 CI + clippy 125 条清零转 gate + lt-app 组织（439+7 测全绿）
- 下载会话线程经 Supervisor::spawn（ThreadRole::Download, Never）：新增
  is_thread_finished 查询（JoinHandle 轮询的监督器版语义），DownloadManager
  在途判定不再直持句柄；worker join panic 转 DownloadFailed 终态（消除
  expect），session catch_unwind 兜底——下载线程 panic 不再永久卡死下载卡
- scripts/check_deps.ps1：§3.1 白名单断言（10 crate 21 条边，deps/dev-deps 双段）
- scripts/check_guards.ps1：§6.2 五组禁令（裸 spawn/直发 proxy/字符串协议/
  panic hook 越位/契约旁路），白名单按实测编目（I/O 泵/wasapi 读环/bench
  内层/下载 worker 子线程为注记例外，tests 夹具豁免）
- ci.yml：两守护脚本进 gate + clippy 转 `-D warnings`——基线 125 条全量
  机械清零（field reassign/map_or/checked ops/chunks as_chunks/doc 格式等）
- lt-app：W5 下沉三路助手拆入 shell_helpers.rs（shell.rs 612→512 行）；
  总行数 1181，验收口径定案 ≤1200（组合根+两项宿主职责的自然结果）

+++ crates/lt-orchestrator/src/logging.rs
@@ -99,7 +99,7 @@ pub fn spawn_bridge(
             match rx.try_recv() {
                 Ok(ev) => artery.push(ev),
                 Err(tokio::sync::broadcast::error::TryRecvError::Lagged(n)) => {
-                    LAG_PENDING.fetch_add(n as u64, Ordering::Relaxed);
+                    LAG_PENDING.fetch_add(n, Ordering::Relaxed);
                     if lag_report_due() {
                         let total = LAG_PENDING.swap(0, Ordering::Relaxed);
                         tracing::debug!(target: BRIDGE_TARGET, "日志桥接丢弃 {total} 行（订阅端积压）");
